fix: #125 Considering reduced dimension of bin
This fixes the problem described in the #125. Instead of creating a
block of 42x42 on top of each individual base and then cutting it, I
create a block based on the grid given and apply an offset and then a
fillet to consider the rounded corners. This produces a geometrically
compatible object, without the problem that was given the other
approach.

I also modified the internal radius constant, so that the wall thickness
on the corner is constant.

In my limited testings, this has been working quite good.