fix: modify place_by_grid to accept size of the grid
We cannot always use the bounding box of the object to get the needed
size for the grid. In the case of the bin base, these are 41.5mm in
size, but the grid still needs to be of 42mm. I added the optional
parameters width and lenght to accomodate this possibility. It does not
introduce a hard dependency for gridfinity, as the values for the width
and length are defined when calling the function. If no values for width
and lenght are added, then the bounding box is used.