test: BaseBlockTest passing
The tests were adjusted to expect results according to the new behaviour
of BaseBlock. In this case the expected results changed more
significantly compared to other tests. This is because the geometry of
BaseBlock changed more significantly. The width and length of the base
block are now 41.5 and 41.5, respectively (from 42.0). Also, the height
is smaller, because it does not have the platform on top any more.