test: GridfinityRefinedMagnetHoleSideTest now passing
After modifying the GridfinityRefinedMagnetHoleSideTest, the expected
results for the test needed to be adjusted. The new values are only
marginally different comapared to the old values. The difference is owed
to the 0.2 mm extra length in one dimension, needed for the object to
cut through the outer wall of the bin.